+++ src/Groups/Groups.js
@@ -10,6 +10,7 @@ import { BatchAllocations } from '../components/BatchAllocations';
 import { GroupsListContainer } from './GroupsList';
 import { CreateGroup } from './CreateGroup';
 import { EditGroup } from './EditGroup';
+import { BatchAllocations } from '../components/BatchAllocations';
 
 const Groups = () => {
   return (
@@ -43,6 +44,14 @@ const Groups = () => {
         render={() => (
           <CheckPermission perm="ui-finance.group.view">
             <GroupsListContainer />
+          </CheckPermission>
+        )}
+      />
+      <Route
+        path={`${GROUPS_ROUTE}/:id/batch-allocations`}
+        render={() => (
+          <CheckPermission perm="ui-finance.allocations.create">
+            <BatchAllocations />
           </CheckPermission>
         )}
       />