+++ src/Ledger/Ledger.js
@@ -21,6 +21,7 @@ import { LedgerListContainer } from './LedgerList';
 import RolloverLedgerContainer, { RolloverLedgerView } from './RolloverLedger';
 import RolloverLedgerCreateFiscalYear from './RolloverLedgerCreateFiscalYear';
 import { RolloverLogs } from './RolloverLogs';
+import { BatchAllocations } from '../components/BatchAllocations';
 
 const Ledger = () => {
   return (