+++ src/common/components/AllocationToolsMenuSection/AllocationToolsMenuSection.js
@@ -13,6 +13,7 @@ import {
 
 export const AllocationToolsMenuSection = ({
   disabled,
+  onBatchAllocate,
   onDownloadAllocationWorksheet,
   onUploadAllocationWorksheet,
 }) => {
@@ -23,6 +24,21 @@ export const AllocationToolsMenuSection = ({
       id="allocation-tools-menu-section"
       label={intl.formatMessage({ id: 'ui-finance.menuSection.allocationTools' })}
     >
+      <IfPermission perm="ui-finance.allocations.create">
+        <Button
+          buttonStyle="dropdownItem"
+          onClick={onBatchAllocate}
+          disabled={disabled}
+        >
+          <Icon
+            size="small"
+            icon="edit"
+          >
+            <FormattedMessage id="ui-finance.actions.allocations.batch" />
+          </Icon>
+        </Button>
+      </IfPermission>
+
       <IfPermission perm="finance.finance-data.collection.get">
         <Button
           buttonStyle="dropdownItem"
@@ -58,6 +74,7 @@ export const AllocationToolsMenuSection = ({
 
 AllocationToolsMenuSection.propTypes = {
   disabled: PropTypes.bool,
+  onBatchAllocate: PropTypes.func.isRequired,
   onDownloadAllocationWorksheet: PropTypes.func.isRequired,
   onUploadAllocationWorksheet: PropTypes.func.isRequired,
 };