+++ src/common/components/index.js
@@ -1,3 +1,4 @@
 export { AllocationToolsMenuSection } from './AllocationToolsMenuSection';
+export { BatchAllocationModal } from './BatchAllocationModal';
 export { DownloadAllocationWorksheetModal } from './DownloadAllocationWorksheetModal';
 export { UploadAllocationWorksheetModal } from './UploadAllocationWorksheetModal';